Move WebSocket configuration outside the Bifrost wrapper.
Previously, the configuration was in the
com.vmware.bifrost.bridge.spring.config.WebSocketConfig class and was
adding a hardcoded "/bifrost" endpoint and configuring the simple message
broker with hardcoded "/topic" and "/pub" destination prefixes.

Now, clients of the Bifrost wrapper should provide WebSocket configuration
in their code base. This has the following benefits:

- Customers will be able to use only the Bifrost message bus without
  enabling the STOMP bridge.

- Customers will have full control over the STOMP endpoint(s) path,
  security settings and destination prefixes. For example, currently in
  the ss-base project we define two stomp endpoints with custom allowed
  origins settings.

- Customers will be able to configure the simple message broker and STOMP
  broker relay in the same place.

In order to identify which are the "Bifrost" STOMP destination prefixes, a
new BifrostBridgeConfiguration bean is added. Users of the wrapper will
have to implement the BifrostBridgeConfigurer interface and register their
Bifrost destinations (e.g. "/topic", "/pub") if they want to use the
bridge functionality. This allows the Bifrost wrapper to ignore
non-bifrost subscription requests, e.g. requests to external message
brokers.