Adding support to associate custom attributes to channels
Adding getChannelAttribute and setChannelAttribute APIs to EventBusLowApi. The new APIs can be used to associate custom attributes to channels. For example, during the H5C will use the new APIs to associate the sensitive clientId with recent tasks channels. Also fixing a concurrency issue in the BifrostSubscriptionService. getOpenChannels() and BifrostSubscriptionService.getSubscriptions() APIs.