Update the Sewing Machine fetch script
A major change to the way the Docker image for Sewing Machine is
built has been made as of v2.301 that resulted in a change in the
signature to the `docker run` command. The benefit of the new
change is in saving build time significantly by simply copying the
already built Sewing Machine CLI binaries from the Docker container
rather than building them every time a container is created and
executed.

However, due to the change in the run command signarture, this script
will no longer work with versions below v2.301. In other words,
the new script will be needed for v2.301 onwards.