Adding RabbitMQ MessageBrokerConnector
The new connector can be used to extend Java EventBus by connecting to RabbitMQ message brokers. The RabbitMQ broker is bundled as a separate vmw-bifrost-rabbitmq.jar to avoid adding new dependencies to the core Bifrost jar.
Added a sample spring boot app which connects to a RabbitMQ broker running in a docker container, check the README.md file for more details.