Adding Kafka MessageBrokerConnector
The new connector can be used to extend Java EventBus by connecting to Kafka message brokers. The Kafka connector is bundled as a separate vmw-bifrost-kafka.jar to avoid adding new dependencies to the core Bifrost jar.
Added a sample spring boot app which connects to a Kafka broker running in a docker container. Check the README.md file for more details how to build and run the sample app.