Stringify payload before broker's send method
The 'payload' property of the MessageObject is not guaranteed to be
serializable, which was causing the message broker's sendMessage
method to throw an exception. To be more exact,
SimpleMessageConverter can only accept an input of String, byte[] or
serializable objects. Assuming the payload has to be considered
for use in the UI, the most optimal way to fix the problem is to
convert and send the payload in the string equivalent.