Adding support to specify BridgeChannelMode for channels.
For now there are only two supported modes: REQUESTS_ONLY and REQUESTS_AND_RESPONSES (default).
If REQUESTS_ONLY mode is specified for a given channel, response message sent to it won't be forwarded to the UI.

This mode solves the following problem with sending requests from the UI to external message brokers like RabbitMQ:

In the TS when we mark a channel as galactic this triggers a “SUBSCRIBE” STOMP command.
When we process the subscribe command in the Java Bifrost, the BifrostSubscriptionService will start listening on the channel response stream in order to forward response messages back to the UI.
However, when the Java channel is marked as galactic listening to its response stream will create new subscription to the external RabbitMQ server. This will be a problem if we want to use the channel only to send requests.