Adding support for store sync between the Bifrost TS and Bifrost Java
Adding new GalacticStoreService which handles store sync requests from the Bifrost TS, the supported client requests are:
- openStore – subscribe the client for store updates for a given store.
- updateStore – update or remove a store item from a given store and notify all subscribed clients for the change.
- closeStore – unsubscribe the client for store updates for a given store.

Added setValueType() and setKeyType() methods to the BusStore interface to specify the store’s key and value types. These types will be used at runtime to deserialize store items coming from the Bifrost TS.
The new methods should be called on stores with non-string keys/values for which we want to enable the store sync feature. Otherwise the items will be deserialized as Map<String, Object>.