Add another overload to sendError()
Previously when private channel support was introduced, an overload to
the sendError method that accepts a Response object as the first
argument was added to support sending Response objects across queues
along with topics. However, there is another method overload that
accepts a generic type that extends GeneralError as the first argument.
This method was overlooked by the time the private channel support was
added, thus limiting the sendError method with the GeneralError variant
to topic-based channels only. This commit will add the missing overload
to fix the issue.